refactor: simplify Markdown semantic updates [ci]
