+++ src/MarkdownSemanticDocument.cs
@@ -1,59 +1,44 @@
-using System.Diagnostics;
 using ICSharpCode.AvalonEdit.Document;
 
 namespace PaperTodo;
 
 /// <summary>
-/// Per-editor semantic cache. TextDocument changes first attempt the existing bounded segmented
-/// incremental parse on the editing thread so presentation can observe current semantics before the
-/// next WPF render. Changes that cannot be handled by that fast path publish immutable source
-/// snapshots to one background worker. At most one background Markdig parse runs at a time; edits
-/// that arrive while it is busy replace a single pending source instead of starting overlapping work.
+/// Per-editor semantic cache owned by the same thread as AvalonEdit's TextDocument. The initial
+/// document and every completed text change synchronously publish exact Markdig semantics before
+/// control returns to WPF rendering. Ordinary edits use the bounded incremental parser first;
+/// edits that cannot be proven local fall back to a full parse on the same thread.
+///
+/// The previous source is retained only for the duration of one edit transaction as AvalonEdit's
+/// cheap immutable Rope snapshot. No permanent worker, semaphore, generation queue or duplicate
+/// full-source string is kept while the editor is idle.
 /// </summary>
 internal sealed class MarkdownSemanticDocument : IDisposable
 {
     private readonly TextDocument _document;
-    private readonly object _gate = new();
-    private readonly SemaphoreSlim _workSignal = new(0, 1);
-    private readonly CancellationTokenSource _disposeCancellation = new();
-    private readonly Task _workerTask;
-    private ITextSource? _pendingSource;
-    private int _pendingGeneration;
-    private MarkdownSemanticSnapshot? _snapshot;
-    private string? _snapshotSource;
-    private int _generation;
-    private int _snapshotGeneration = -1;
-    private int _earliestChangedOffsetSincePublished = int.MaxValue;
-    private bool _lineStructureChangedSincePublished;
+    private MarkdownSemanticSnapshot _snapshot;
+    private ITextSource? _sourceBeforeChange;
     private bool _disposed;
 
     public MarkdownSemanticDocument(TextDocument document)
     {
         _document = document ?? throw new ArgumentNullException(nameof(document));
-        _document.Changed += OnDocumentChanged;
+        _snapshot = MarkdownSemanticSnapshot.Parse(_document.Text);
+        _document.Changing += OnDocumentChanging;
         _document.TextChanged += OnDocumentTextChanged;
-        _workerTask = Task.Run(ProcessQueueAsync);
-        QueueCurrentSource();
     }
 
     /// <summary>
-    /// Raised after the semantic snapshot for the current source generation is published. The event
-    /// may be raised either from the editing thread's incremental fast path or from the parsing worker;
-    /// WPF consumers must marshal through their Dispatcher.
+    /// Raised synchronously after semantics for the completed TextDocument change are published.
+    /// Consumers may still defer visual invalidation to their normal WPF render priority.
     /// </summary>
     public event Action? SnapshotChanged;
 
     public bool TryGetCurrent(out MarkdownSemanticSnapshot snapshot)
     {
-        lock (_gate)
+        if (!_disposed)
         {
-            if (!_disposed &&
-                _snapshot != null &&
-                _snapshotGeneration == _generation)
-            {
-                snapshot = _snapshot;
-                return true;
-            }
+            snapshot = _snapshot;
+            return true;
         }
 
         snapshot = null!;
@@ -61,295 +46,92 @@ internal sealed class MarkdownSemanticDocument : IDisposable
     }
 
     /// <summary>
-    /// Returns the most recently published snapshot even when a newer source generation is pending.
-    /// Dirty-range metadata lets editor helpers use stale semantics only when the source range they
-    /// care about is provably before every change since that publication.
+    /// Compatibility surface for editor helpers that previously had to reason about an async stale
+    /// snapshot. Publication is now synchronous, so a live document's latest snapshot is always
+    /// current and there is no pending dirty range.
     /// </summary>
     public bool TryGetLatest(
         out MarkdownSemanticSnapshot snapshot,
         out int earliestChangedOffset,
         out bool lineStructureChanged)
     {
-        lock (_gate)
+        if (!_disposed)
         {
-            if (_disposed || _snapshot == null)
-            {
-                snapshot = null!;
-                earliestChangedOffset = 0;
-                lineStructureChanged = true;
-                return false;
-            }
-
             snapshot = _snapshot;
-            earliestChangedOffset = _snapshotGeneration == _generation
-                ? int.MaxValue
-                : _earliestChangedOffsetSincePublished;
-            lineStructureChanged = _snapshotGeneration != _generation &&
-                _lineStructureChangedSincePublished;
+            earliestChangedOffset = int.MaxValue;
+            lineStructureChanged = false;
             return true;
         }
+
+        snapshot = null!;
+        earliestChangedOffset = 0;
+        lineStructureChanged = true;
+        return false;
     }
 
-    /// <summary>
-    /// Forces exact current semantics. Keep this for deliberate click/command paths only; ordinary
-    /// presentation, pointer hover and plain typing must not call it.
-    /// </summary>
     public MarkdownSemanticSnapshot Current()
     {
-        while (true)
-        {
-            if (TryGetCurrent(out var current))
-            {
-                return current;
-            }
-
-            int generation;
-            MarkdownSemanticSnapshot? baseSnapshot;
-            string? baseSource;
-            lock (_gate)
-            {
-                ObjectDisposedException.ThrowIf(_disposed, this);
-                generation = _generation;
-                baseSnapshot = _snapshot;
-                baseSource = _snapshotSource;
-            }
-
-            // CreateSnapshot is the one TextDocument operation explicitly safe for immutable
-            // cross-thread consumption. Materialize its string only for the actual parse.
-            var source = _document.CreateSnapshot().Text;
-            var parsed = ParseAgainstPublished(source, baseSnapshot, baseSource);
-            if (TryPublish(generation, source, parsed))
-            {
-                return parsed;
-            }
-        }
+        ObjectDisposedException.ThrowIf(_disposed, this);
+        return _snapshot;
     }
 
-    private void OnDocumentChanged(object? sender, DocumentChangeEventArgs e)
+    private void OnDocumentChanging(object? sender, DocumentChangeEventArgs e)
     {
-        lock (_gate)
+        if (_disposed || _sourceBeforeChange != null)
         {
-            if (_disposed)
-            {
-                return;
-            }
-
-            _generation++;
-            _earliestChangedOffsetSincePublished = Math.Min(
-                _earliestChangedOffsetSincePublished,
-                e.Offset);
-            _lineStructureChangedSincePublished |=
-                ContainsLineBreak(e.InsertedText) ||
-                ContainsLineBreak(e.RemovedText);
+            return;
         }
+
+        // TextDocument may batch several primitive mutations inside BeginUpdate/EndUpdate. Capture
+        // the source only before the first mutation so the semantic snapshot and old source describe
+        // the same complete pre-edit generation. RopeTextSource shares immutable rope nodes and does
+        // not materialize another whole-document string here.
+        _sourceBeforeChange = _document.CreateSnapshot();
     }
 
     private void OnDocumentTextChanged(object? sender, EventArgs e)
     {
-        // AvalonEdit invalidates the edited visual line immediately. If semantics remain pending
-        // until the worker publishes, the line can render once without heading/code metrics and then
-        // render again with them, which is visible as a one-frame input wobble. The production
-        // segmented parser is already bounded to a small adaptive window for ordinary edits, so try
-        // that path synchronously and keep the worker as the fallback for expensive/global cases.
-        if (!TryPublishIncrementalCurrentSource())
+        if (_disposed)
         {
-            QueueCurrentSource();
-        }
-    }
-
-    private bool TryPublishIncrementalCurrentSource()
-    {
-        int generation;
-        MarkdownSemanticSnapshot? baseSnapshot;
-        string? baseSource;
-        lock (_gate)
-        {
-            if (_disposed || _snapshot == null || _snapshotSource == null)
-            {
-                return false;
-            }
-
-            generation = _generation;
-            baseSnapshot = _snapshot;
-            baseSource = _snapshotSource;
+            return;
         }
 
-        var source = _document.CreateSnapshot().Text;
-        if (!MarkdownSemanticSnapshot.TryParseSegmentedIncremental(
-                baseSource,
-                baseSnapshot,
+        var sourceBeforeChange = _sourceBeforeChange;
+        _sourceBeforeChange = null;
+        var source = _document.Text;
+
+        MarkdownSemanticSnapshot next;
+        if (sourceBeforeChange != null &&
+            MarkdownSemanticSnapshot.TryParseIncremental(
+                sourceBeforeChange.Text,
+                _snapshot,
                 source,
                 out var incremental,
                 out _))
         {
-            return false;
+            next = incremental;
         }
-
-        return TryPublish(generation, source, incremental);
-    }
-
-    private void QueueCurrentSource()
-    {
-        var source = _document.CreateSnapshot();
-        lock (_gate)
+        else
         {
-            if (_disposed)
-            {
-                return;
-            }
-
-            _pendingSource = source;
-            _pendingGeneration = _generation;
+            next = MarkdownSemanticSnapshot.Parse(source);
         }
 
-        SignalWorker();
-    }
-
-    private void SignalWorker()
-    {
-        try
-        {
-            if (_workSignal.CurrentCount == 0)
-            {
-                _workSignal.Release();
-            }
-        }
-        catch (SemaphoreFullException)
-        {
-        }
-    }
-
-    private async Task ProcessQueueAsync()
-    {
-        var cancellationToken = _disposeCancellation.Token;
-        while (true)
-        {
-            try
-            {
-                await _workSignal.WaitAsync(cancellationToken).ConfigureAwait(false);
-            }
-            catch (OperationCanceledException)
-            {
-                return;
-            }
-
-            ITextSource? source;
-            int generation;
-            MarkdownSemanticSnapshot? baseSnapshot;
-            string? baseSource;
-            lock (_gate)
-            {
-                if (_disposed)
-                {
-                    return;
-                }
-
-                source = _pendingSource;
-                generation = _pendingGeneration;
-                _pendingSource = null;
-                baseSnapshot = _snapshot;
-                baseSource = _snapshotSource;
-            }
-
-            if (source == null)
-            {
-                continue;
-            }
-
-            try
-            {
-                var sourceText = source.Text;
-                var parsed = ParseAgainstPublished(sourceText, baseSnapshot, baseSource);
-                TryPublish(generation, sourceText, parsed);
-            }
-            catch (Exception ex)
-            {
-                Trace.TraceError("Markdown semantic background parse failed: {0}", ex);
-            }
-
-            // Changes arriving during Parse leave one signal and one latest pending source. The next
-            // loop consumes exactly that newest source; obsolete intermediate generations disappear.
-        }
-    }
-
-    private static MarkdownSemanticSnapshot ParseAgainstPublished(
-        string source,
-        MarkdownSemanticSnapshot? baseSnapshot,
-        string? baseSource)
-    {
-        if (baseSnapshot != null &&
-            baseSource != null &&
-            MarkdownSemanticSnapshot.TryParseSegmentedIncremental(
-                baseSource,
-                baseSnapshot,
-                source,
-                out var incremental,
-                out _))
-        {
-            return incremental;
-        }
-
-        return MarkdownSemanticSnapshot.Parse(source);
-    }
-
-    private bool TryPublish(int generation, string source, MarkdownSemanticSnapshot snapshot)
-    {
-        Action? changed;
-        lock (_gate)
-        {
-            if (_disposed || generation != _generation)
-            {
-                return false;
-            }
-
-            if (_snapshot != null && _snapshotGeneration == generation)
-            {
-                return true;
-            }
-
-            _snapshot = snapshot;
-            _snapshotSource = source;
-            _snapshotGeneration = generation;
-            _earliestChangedOffsetSincePublished = int.MaxValue;
-            _lineStructureChangedSincePublished = false;
-            changed = SnapshotChanged;
-        }
-
-        changed?.Invoke();
-        return true;
-    }
-
-    private static bool ContainsLineBreak(ITextSource text)
-    {
-        if (text.TextLength <= 0)
-        {
-            return false;
-        }
-
-        return text.IndexOf('\r', 0, text.TextLength) >= 0 ||
-            text.IndexOf('\n', 0, text.TextLength) >= 0;
+        _snapshot = next;
+        SnapshotChanged?.Invoke();
     }
 
     public void Dispose()
     {
-        lock (_gate)
+        if (_disposed)
         {
-            if (_disposed)
-            {
-                return;
-            }
-
-            _disposed = true;
-            _document.Changed -= OnDocumentChanged;
-            _document.TextChanged -= OnDocumentTextChanged;
-            _pendingSource = null;
-            _snapshot = null;
-            _snapshotSource = null;
-            _snapshotGeneration = -1;
-            SnapshotChanged = null;
+            return;
         }
 
-        _disposeCancellation.Cancel();
-        _disposeCancellation.Dispose();
+        _disposed = true;
+        _document.Changing -= OnDocumentChanging;
+        _document.TextChanged -= OnDocumentTextChanged;
+        _sourceBeforeChange = null;
+        _snapshot = MarkdownSemanticSnapshot.Empty;
+        SnapshotChanged = null;
     }
 }
