+++ tests/PaperTodo.MarkdownSemanticChecks/PaperTodo.MarkdownSemanticChecks.csproj
@@ -15,7 +15,6 @@
     <Compile Include="..\..\src\MarkdownSemanticSnapshot.Escapes.cs" Link="MarkdownSemanticSnapshot.Escapes.cs" />
     <Compile Include="..\..\src\MarkdownSemanticSnapshot.Compatibility.cs" Link="MarkdownSemanticSnapshot.Compatibility.cs" />
     <Compile Include="..\..\src\MarkdownSemanticSnapshot.Incremental.cs" Link="MarkdownSemanticSnapshot.Incremental.cs" />
-    <Compile Include="..\..\src\MarkdownSemanticSnapshot.Segments.cs" Link="MarkdownSemanticSnapshot.Segments.cs" />
     <Compile Include="..\..\src\MarkdownListEditing.cs" Link="MarkdownListEditing.cs" />
     <Compile Include="..\..\src\MarkdownImageReferences.cs" Link="MarkdownImageReferences.cs" />
   </ItemGroup>
